end_OK

+++ resources/views/quizzes/show.blade.php
@@ -7,25 +7,31 @@
 
             {{  $quiz->title }}
 
-            <x-slot name="right">
-                <x-button-link href="{{ route('quizzes.edit', $quiz->id) }}">
-                    {{__('Редактировать')}}
-                </x-button-link>
-            </x-slot>
+            @auth
+                @if(auth()->user()->isAdmin()) <!-- Проверьте, что метод isAdmin() верно определен в вашей модели User -->
+                    <x-slot name="right">
+                        <x-button-link href="{{ route('quizzes.edit', $quiz->id) }}">
+                            {{__('Редактировать')}}
+                        </x-button-link>
+                    </x-slot>
+        
+                    <x-slot name="delete">
+                        <x-form action="{{route('quizzes.delete', $quiz->id)}} " method="delete">
+                            <x-button type="submit" size="sm">
+                                {{__('Удалить')}}
+                            </x-button-link>
+                        </x-form>
+                    </x-slot>
+        
+                    <x-slot name="question">
+                        <x-button-link href="{{ route('questions.create', $quiz->id) }}">
+                            {{__('Добавить вопрос')}}
+                        </x-button-link>
+                    </x-slot>
+                @endif
+            @endauth
 
-            <x-slot name="delete">
-                <x-form action="{{route('quizzes.delete', $quiz->id)}} " method="delete">
-                    <x-button type="submit" size="sm">
-                        {{__('Удалить')}}
-                    </x-button-link>
-                </x-form>
-            </x-slot>
-
-            <x-slot name="question">
-                <x-button-link href="{{ route('questions.create', $quiz->id) }}">
-                    {{__('Добавить вопрос')}}
-                </x-button-link>
-            </x-slot>
+            
   
 
         </x-title>
